♻ refactor: repository 분리
