+++ OneCoin/Server/src/main/java/OneCoin/Server/upbit/websocket/publisher/UpbitWebSocketPublisher.java
@@ -4,6 +4,8 @@ import OneCoin.Server.dto.SingleResponseDto;
 import OneCoin.Server.upbit.dto.MarketDto;
 import OneCoin.Server.upbit.dto.orderbook.OrderBookDto;
 import OneCoin.Server.upbit.dto.ticker.TickerDto;
+import OneCoin.Server.upbit.repository.OrderBookRepository;
+import OneCoin.Server.upbit.repository.TickerRepository;
 import OneCoin.Server.upbit.service.UpbitHandlingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
@@ -16,12 +18,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UpbitWebSocketPublisher {
     private final SimpMessagingTemplate messagingTemplate;
-    private final UpbitHandlingService upbitHandlingService;
+    private final TickerRepository tickerRepository;
+    private final OrderBookRepository orderBookRepository;
 
     @Scheduled(fixedDelay = 1000)
     public void send() {
-        List<TickerDto> tickerDto = upbitHandlingService.findTickers();
-        List<OrderBookDto> orderBookDto = upbitHandlingService.findOrderBooks();
+        List<TickerDto> tickerDto = tickerRepository.findTickers();
+        List<OrderBookDto> orderBookDto = orderBookRepository.findOrderBooks();
         MarketDto marketDto = new MarketDto(tickerDto, orderBookDto);
         messagingTemplate.convertAndSend("/info/upbit", new SingleResponseDto<>(marketDto));
     }
